src/operations.py
```python
import cloud
from src import event_management, recommendation_system, virtual_networking, analytics, automation_tools, customer_support

class Operations:
    def __init__(self):
        self.event_management = event_management.EventManagement()
        self.recommendation_system = recommendation_system.RecommendationSystem()
        self.virtual_networking = virtual_networking.VirtualNetworking()
        self.analytics = analytics.Analytics()
        self.automation_tools = automation_tools.AutomationTools()
        self.customer_support = customer_support.CustomerSupport()

    def manage_event(self, eventData):
        self.event_management.registerUser(eventData)
        self.event_management.purchaseTicket(eventData)
        self.event_management.scheduleEvent(eventData)

    def provide_recommendations(self, userData):
        self.recommendation_system.recommendSessions(userData)
        self.recommendation_system.recommendSpeakers(userData)
        self.recommendation_system.recommendNetworkingOpportunities(userData)

    def enable_networking(self, userData):
        self.virtual_networking.startLiveChat(userData)
        self.virtual_networking.createBreakoutRoom(userData)
        self.virtual_networking.network(userData)

    def track_engagement(self, userData, eventData):
        self.analytics.trackEngagement(userData, eventData)

    def automate_operations(self, eventData):
        self.automation_tools.automateRegistration(eventData)
        self.automation_tools.automateTicketing(eventData)
        self.automation_tools.automateScheduling(eventData)

    def provide_support(self, userData):
        self.customer_support.resolveIssue(userData)
        self.customer_support.provideUpdates(userData)

    def run_operations(self, userData, eventData):
        self.manage_event(eventData)
        self.provide_recommendations(userData)
        self.enable_networking(userData)
        self.track_engagement(userData, eventData)
        self.automate_operations(eventData)
        self.provide_support(userData)
```