tests/test_operations.py
```python
import unittest
from src import operations

class TestOperations(unittest.TestCase):

    def setUp(self):
        self.operations = operations.Operations()
        self.userData = {"name": "John Doe", "email": "johndoe@example.com"}
        self.eventData = {"name": "Virtual Concert", "date": "2022-12-31", "tickets": 100}

    def test_registerUser(self):
        result = self.operations.registerUser(self.userData)
        self.assertEqual(result, "User registered successfully")

    def test_purchaseTicket(self):
        result = self.operations.purchaseTicket(self.userData, self.eventData)
        self.assertEqual(result, "Ticket purchased successfully")

    def test_scheduleEvent(self):
        result = self.operations.scheduleEvent(self.eventData)
        self.assertEqual(result, "Event scheduled successfully")

    def test_startLiveChat(self):
        result = self.operations.startLiveChat(self.userData)
        self.assertEqual(result, "Live chat started")

    def test_createBreakoutRoom(self):
        result = self.operations.createBreakoutRoom(self.userData)
        self.assertEqual(result, "Breakout room created")

    def test_network(self):
        result = self.operations.network(self.userData)
        self.assertEqual(result, "Networking started")

    def test_trackEngagement(self):
        result = self.operations.trackEngagement(self.userData, self.eventData)
        self.assertEqual(result, "Engagement tracked")

    def test_resolveIssue(self):
        result = self.operations.resolveIssue(self.userData, "Ticketing issue")
        self.assertEqual(result, "Issue resolved")

if __name__ == '__main__':
    unittest.main()
```